Use u32 instead of u8 in many places
An 8-bit integer was used in many places because it was sufficient for
the required range. But if it is not going to be stored in a collection,
space is not really an issue, and CPUs can do arithmetic with 32-bit
integers more efficiently than 8-bit integers, so use u32 where
appropriate.

+++ src/frame.rs
@@ -46,7 +46,7 @@ struct FrameHeader {
     pub block_size: u16,
     pub sample_rate: Option<u32>,
     pub channel_assignment: ChannelAssignment,
-    pub bits_per_sample: Option<u8>,
+    pub bits_per_sample: Option<u32>,
 }
 
 impl FrameHeader {
